Extracted attachment tests into a separate file.
