+++ example/example_test.go
@@ -2,8 +2,6 @@ package example_test
 
 import (
 	"github.com/dailymotion/allure-go/example"
-	"io/ioutil"
-	"log"
 	"testing"
 
 	"github.com/dailymotion/allure-go"
@@ -67,41 +65,5 @@ func TestFailedExample(t *testing.T) {
 func TestSkippedExample(t *testing.T) {
 	allure.Test(t, "This is a test to show allure implementation with a skipped test", func() {
 		t.Skip("Skipping this test")
-	})
-}
-
-func TestTextAttachmentToStep(t *testing.T) {
-	allure.Test(t, "Testing a text attachment", func() {
-		allure.Step("adding a text attachment", func() {
-			_ = allure.AddTextAttachment("text!", allure.TextPlain, "Some text!")
-		})
-	})
-}
-
-func TestImageAttachmentToStep(t *testing.T) {
-	allure.Test(t, "testing an image attachment", func() {
-		allure.Step("adding an image attachment", func() {
-			dat, err := ioutil.ReadFile("../Coryphaena_hippurus.png")
-			if err != nil {
-				log.Fatal(err)
-			}
-			_ = allure.AddByteArrayAttachment("mahi-mahi", allure.ImagePng, dat)
-		})
-	})
-}
-
-func TestTextAttachment(t *testing.T) {
-	allure.Test(t, "Testing a text attachment", func() {
-		_ = allure.AddTextAttachment("text!", allure.TextPlain, "Some text!")
-	})
-}
-
-func TestImageAttachment(t *testing.T) {
-	allure.Test(t, "testing an image attachment", func() {
-		dat, err := ioutil.ReadFile("../Coryphaena_hippurus.png")
-		if err != nil {
-			log.Fatal(err)
-		}
-		_ = allure.AddByteArrayAttachment("mahi-mahi", allure.ImagePng, dat)
 	})
 }
